+++ src/flows/CostCalculator/utils.ts
@@ -39,6 +39,7 @@ export function buildPayload(
     employer_currency_slug: values.currency,
     include_benefits: estimationOptions.includeBenefits,
     include_cost_breakdowns: estimationOptions.includeCostBreakdowns,
+    include_premium_benefits: estimationOptions.includePremiumBenefits,
     employments: [
       {
         region_slug: values.region || values.country,