+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # @remoteoss/remote-flows
 
+## 0.0.4
+
+### Patch Changes
+
+- 48544df: Add `CostCalculatorFlow` and related components to build the CostCalculator in a more flexible way
+- 8af7ae7: Assign label as placeholders
+
+## 0.0.3
+
+### Patch Changes
+
+- 2eb1183: - Add include_premium_benefits to the cost calculator flow
+- ce8dec3: - Export useCostCalculator hook to consumers
+
 ## 0.0.2
 
 ### Patch Changes